# Performance: replace a slow numpy function