@Cregrant 2023-02-07
The second commit will speed up the loading of aligned faces.
Test data: 14500 samples, 676 MB

Single thread loading (reference): 61s
```
        result = []
        for filename in image_paths:
            result.append(load_face_sample(filename))
```

Custom subprocess loading (DFL current): 58s
```
        result = FaceSamplesLoaderSubprocessor(image_paths).run()
```

multiprocessing.Pool() loading (commit): 18s
```
        with multiprocessing.Pool() as pool:
            result = pool.map(load_face_sample, image_paths)
```