@iperov 2023-02-08
it's not fast. 
```out=``` does in-place modification breaking functional style