@Cregrant 2023-02-08
You can remove `out=` in that case. It still faster than a numpy clip